+++ src/main/kotlin/lotto/domain/LottoStatistics.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoStatistics(
     private val lottoTickets: List<LottoTicket>,
