+++ src/main/kotlin/lotto/domain/LottoTicketMachine.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 object LottoTicketMachine {
